Abstract route schemas to a separate module
